backend/requirements.txt
fastapi
uvicorn
pandas
qdrant-client
sentence-transformers
python-dotenv
openpyxl
